Refactor: Unify target value retrieval and simplify health ViewModels
This commit introduces a significant refactoring to centralize the logic for retrieving profile-specific target values (e.g., target weight, steps, sleep) and simplifies the state management within several health-related ViewModels.

A new reusable function, `getTargetValueFlow`, has been added to `ProfileRepository` to abstract away the boilerplate of observing the current profile and extracting a target value. This cleans up multiple ViewModels and reduces code duplication.

Additionally, the `WeightOverview` components have been renamed to `BodyWeightOverview` for clarity and consistency.

### Key Changes:

-   **`ProfileRepository.kt`:**
    -   Introduced a new generic function `getTargetValueFlow<E: Number>(...)` that takes a lambda to extract a target value from the current user's profile.
    -   This function encapsulates the logic of listening to `settingsRepository` for the selected profile ID and then fetching the corresponding profile data.

-   **ViewModel Refactoring (Target Value Logic):**
    -   Removed duplicated `flatMapLatest` logic for fetching target values from `BeverageDailyViewModel`, `BeverageOverviewViewModel`, `SleepDailyViewModel`, `SleepOverviewViewModel`, `StepsDailyViewModel`, `StepsOverviewViewModel`, `BodyWeightDailyViewModel`, and `BodyWeightOverviewViewModel`.
    -   All these ViewModels now use the new `profileRepository.getTargetValueFlow` for a cleaner, single-line implementation.

-   **State Management Simplification:**
    -   **`BodyWeightOverviewViewModel` & `BodyWeightDailyViewModel`:**
        -   Removed `maxWeight` and `targetWeight` as separate `StateFlow`s collected in the View.
        -   These values are now directly included in the `BodyWeightOverviewUiState` and `BodyWeightDailyUiState`, simplifying the composables (`BodyWeightOverviewView`, `BodyWeightDailyView`).
        -   The logic to calculate `maxBodyWeight` is now self-contained within the ViewModel's `combine` block.
    -   **Other Health ViewModels (`Steps`, `Sleep`, `Beverage`):**
        -   Similarly, `maxSteps`, `maxSleeps`, and `maxBeverages` are now part of their respective `UiState` data classes. This removes the need for views to collect multiple state flows from the ViewModel.
    -   The calculation for the maximum value for chart scaling now correctly uses `roundUpToNextTen()` for better visualization.

-   **Renaming for Clarity:**
    -   All components related to weight overview have been renamed for consistency:
        -   `WeightOverviewViewModel` -> `BodyWeightOverviewViewModel`
        -   `WeightOverviewView` -> `BodyWeightOverviewView`
        -   `WeightOverviewUiState` -> `BodyWeightOverviewUiState`
        -   `WeightOverviewError` -> `BodyWeightOverviewError`
    -   This change is reflected across all call sites, including `BottomBarNavGraph.kt`, `Modules.kt`, and `HealthOverviewViewModel.kt`.

-   **UI & Plotting Adjustments:**
    -   Composables like `PlotBodyWeights`, `PlotBeverages`, and their respective views have been updated to consume `max` and `target` values directly from the `UiState` object, leading to cleaner function signatures and view logic.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/beverages/PlotBeverages.kt
@@ -51,9 +51,9 @@ fun PlotBeverages(
     beverages: List<BeverageOverview>,
     dateRange: DateRange,
     dates: List<LocalDate>,
-    maxBeverages: UInt,
+    maxBeverages: Int,
     thumbnail: Boolean = false,
-    targetBeverages: UInt? = null
+    targetBeverages: Int? = null
 ) {
     ChartLayout(
         modifier = modifier.padding(horizontal = 8.dp),
@@ -67,7 +67,7 @@ fun PlotBeverages(
 
         XYGraph(
             xAxisModel = CategoryAxisModel(dates),
-            yAxisModel = IntLinearAxisModel(0..maxBeverages.toInt()),
+            yAxisModel = IntLinearAxisModel(0..maxBeverages),
             horizontalMajorGridLineStyle = null,
             horizontalMinorGridLineStyle = null,
             verticalMajorGridLineStyle = null,
@@ -172,9 +172,9 @@ fun PlotBeverages(
                 )
             }
 
-            if (targetBeverages != null && targetBeverages > 0u) {
+            if (targetBeverages != null && targetBeverages > 0) {
                 LinePlot(
-                    data = dates.map { Point(it, targetBeverages.toInt()) },
+                    data = dates.map { Point(it, targetBeverages) },
                     lineStyle =
                         LineStyle(
                             brush = SolidColor(Color(0xFFED7D31)),
